fix bash completions error after installation
After a clean `dotly` installation, when running `bash`, the following error is shown in the shell:

`bash: /home/dani/.dotfiles/shell/bash/completions/*: No such file or directory`

The root cause of this error is the following: when it searches for bash completions, it takes hidden files (ls -A) into account. Since the `.gitkeep` file exists in the `$DOTFILES_PATH/shell/bash/completions/` directory, it tries to load all completion files from that directory, but since it is empty (it ignores hidden files when traversing all files in the directory) we get the above error.