Fix(import): Error importing packages
Solved an error when importing packages from a previous machine if the file name is similar but not exactly the same the import ocurred with a different name. For example if your brewfile (or any other) previous name was "this-is-my-computer-old" and the new name for your computer was "this-is-my-computer". The script were importing "this-is-my-computer-old".